+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning].
     - Remove the `Maybe<T>` struct.
     - Remove tuple extension methods.
     - Remove `IResult` and `IResult<T>` interfaces.
+    - Remove "replace error" feature in `FailResult`.
 
 ## [1.4.0] - 2023-06-28